% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insilico.R
\name{extract}
\alias{extract}
\title{Extract amplicons from sequences from \code{insilicoPCR} object}
\usage{
extract(x)
}
\arguments{
\item{x}{object of class \code{insilicoPCR}}
}
\value{
object of class \code{DNAStringSet}
}
\description{
Extract amplicons from sequences from \code{insilicoPCR} object
}
\examples{
##Run local primer blast

query <- readDNAStringSet("ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCA/000/010/525/GCA_000010525.1_ASM1052v1/GCA_000010525.1_ASM1052v1_genomic.fna.gz")

primers <- c("GGCTGGATCACCTCCTT","GCCWAGGCATCCDCC")
db <- makeprimerdb(primers)

#Define BLAST arguments: these are the same that can be found in `RinsilicoPCR::blast.params$primers`
args <- list(
		outfmt="'6 qacc sacc pident length qlen slen qstart sstart qend send evalue'",
		max_target_seqs=20,
		word_size=8
		)

blastres <- blast(query,db,args)

pairs <- primer.insilico(blastres)

amplicons <- extract(pairs)

}
