% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insilico.R
\name{makeprimerdb}
\alias{makeprimerdb}
\title{Create a BLAST Database for Insilico PCR primers}
\usage{
makeprimerdb(x)
}
\arguments{
\item{x}{a vector or list of length 2 containing a forward primer (5'->3' on plus strand) and reverse primer (5'->3' on minus strand)}
}
\value{
String giving the path to the BLAST database
}
\description{
Demultiplexes degenerate bases in a set of forward and reverse primers and creates a blast database
}
\examples{
primers <- c("GGCTGGATCACCTCCTT","GCCWAGGCATCCDCC")
db <- makeprimerdb(primers)



}
