% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insilico.R
\name{primer.insilico}
\alias{primer.insilico}
\title{Find primer pairs in BLAST results}
\usage{
primer.insilico(x, filt = (pident >= 80 & length/slen >= 0.8),
  size.range = c(50, 1200))
}
\arguments{
\item{x}{data.frame of BLAST results}

\item{filt}{Logical expression used to filter mismatches from primer BLAST results. Defaults to `pident >= 80 & length/slen >= .8`, (percent identity > 80\% and match length > 80\% of primer).}

\item{size.range}{numeric vector of length 2 giving the range of expected amplicon length.}
}
\value{
list of results for each query sequence containing:

dir : orientation of amplicon

blastres : BLAST table results for primer pair

location : start and end coordinates for the amplicon

size : size of the amplicon
}
\description{
Find primer pairs in BLAST results
}
\examples{
##Run local primer blast

query <- readDNAStringSet("ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCA/000/010/525/GCA_000010525.1_ASM1052v1/GCA_000010525.1_ASM1052v1_genomic.fna.gz")

primers <- c("GGCTGGATCACCTCCTT","GCCWAGGCATCCDCC")
db <- makeprimerdb(primers)

#Define BLAST arguments: these are the same that can be found in `RinsilicoPCR::blast.params$primers`
args <- list(
		outfmt="'6 qacc sacc pident length qlen slen qstart sstart qend send evalue'",
		max_target_seqs=20,
		word_size=8
		)

blastres <- blast(query,db,args)

pairs <- primer.insilico(blastres)
}
